+++ RecoEgamma/ElectronIdentification/interface/ElectronDNNEstimator.h
@@ -17,22 +17,22 @@ public:
   ;
 
   // Function returning a map with all the possible variables and their name
-  std::map<std::string, float> getInputsVars(const reco::GsfElectron& ele) const;
+  std::map<std::string, double> getInputsVars(const reco::GsfElectron& ele) const;
 
   // Evaluate the DNN on all the electrons with the correct model
-  std::vector<std::vector<float>> evaluate(const reco::GsfElectronCollection& ele,
-                                           const std::vector<tensorflow::Session*>& sessions) const;
+  std::vector<std::vector<double>> evaluate(const reco::GsfElectronCollection& ele,
+                                            const std::vector<tensorflow::Session*>& sessions) const;
 
   // List of input variables names used to check the variables request as
   // inputs in a dynamic way from configuration file.
   // If an input variables is not found at construction time an expection is thrown.
   static const std::vector<std::string> dnnAvaibleInputs;
 
-  static constexpr float ptThreshold = 10.;
-  static constexpr float ecalBarrelMaxEtaWithGap = 1.566;
-  static constexpr float ecalBarrelMaxEtaNoGap = 1.485;
-  static constexpr float endcapBoundary = 2.5;
-  static constexpr float extEtaBoundary = 2.65;
+  static constexpr double ptThreshold = 10.;
+  static constexpr double ecalBarrelMaxEtaWithGap = 1.566;
+  static constexpr double ecalBarrelMaxEtaNoGap = 1.485;
+  static constexpr double endcapBoundary = 2.5;
+  static constexpr double extEtaBoundary = 2.65;
 
 private:
   const egammaTools::EgammaDNNHelper dnnHelper_;
